+++ onnx_synthesize.py
@@ -41,9 +41,9 @@ if __name__ == "__main__":
 
     # Preprocess texts
     ids = raw_texts = [args.text[:100]]
-    speakers = np.array([args.speaker_id], dtype=np.int32)
+    speakers = np.array([args.speaker_id], dtype=np.int64)
     accents = None
-    phonemes, accents = [np.array([out]).astype(dtype=np.int32) for out in preprocess_japanese(args.text)]
+    phonemes, accents = [np.array([out]).astype(dtype=np.int64) for out in preprocess_japanese(args.text)]
 
     variance_session = onnxruntime.InferenceSession("variance_model.onnx", providers=['CUDAExecutionProvider'])
     embedder_session = onnxruntime.InferenceSession("embedder_model.onnx", providers=['CUDAExecutionProvider'])
